Acceptance tests - infrastructure problems
#11, #21, #51
- Race condition in user bridge initialization in the driver (to allow parallelism)
- Decoupled tests requiring a user to be a member from their order of execution (for example, the sign up test does not have to run before those).
The problem arises because the system does not have a function to remove member users from the system.
The solution was to add another method to the bridge interface that only signs a user up if it wasn't already signed up.
This also required those tests to agree on the same username and password.